refactor: rename uiHandlers to domHandlers
