+++ src/webview/index.html
@@ -23,7 +23,7 @@
           "./modules/uiManagers/SettingsButtonManager.js": "${settingsButtonManagerUri}",
           "./modules/uiManagers/RecordingManager.js": "${recordingManagerUri}",
           "./modules/uiManagers/InstructionManager.js": "${instructionManagerUri}",
-          "./modules/uiHandlers.js": "${uiHandlersUri}",
+          "./modules/domHandlers.js": "${domHandlersUri}",
           "./modules/messageHandlers.js": "${messageHandlersUri}",
           "@common/webviewContext.js": "${webviewContextUri}",
           "@common/webviewState.js": "${webviewStateUri}",
